pyproject.toml
[project]
name = "gabut"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncio>=3.4.3",
    "cffi>=1.17.1",
    "chromadb>=1.0.9",
    "cmake>=4.0.2",
    "dataclasses>=0.8",
    "future>=1.0.0",
    "google-genai>=1.16.1",
    "ipykernel>=6.29.5",
    "ipython>=9.2.0",
    "langchain-community>=0.3.24",
    "langchain-google-genai>=2.1.4",
    "langchain-openai>=0.3.18",
    "langgraph>=0.4.5",
    "llama-index>=0.12.37",
    "llama-index-embeddings-huggingface>=0.5.4",
    "llama-index-llms-google-genai>=0.1.13",
    "llama-index-llms-huggingface-api>=0.4.3",
    "llama-index-tools-google>=0.3.0",
    "llama-index-tools-mcp>=0.1.3",
    "llama-index-utils-workflow>=0.3.2",
    "llama-index-vector-stores-chroma>=0.4.1",
    "mkl>=2025.1.0",
    "mkl-include>=2025.1.0",
    "numpy>=2.2.6",
    "opencv-python>=4.11.0.86",
    "paddleocr>=2.0.6",
    "paddlepaddle>=2.0.6",
    "pillow>=11.2.1",
    "pymupdf>=1.26.0",
    "pypdf>=5.5.0",
    "pyside6>=6.9.0",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "setuptools>=80.8.0",
    "six>=1.17.0",
    "smolagents>=1.16.1",
    "torch>=2.7.0",
    "torchaudio>=2.7.0",
    "torchvision>=0.22.0",
    "typing-extensions>=4.13.2",
]
